+++ src/main/java/persistence/sql/ddl/constraint/ColumnH2Constraint.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.constraint;
 
 import jakarta.persistence.Column;
