+++ src/main/java/persistence/sql/ddl/constraint/H2Constraint.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.constraint;
 
 import java.lang.reflect.Field;
