+++ src/main/java/persistence/sql/ddl/entity/Person.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
