+++ src/main/java/persistence/sql/ddl/mapper/H2ConstraintMapper.java
@@ -1,8 +1,12 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.mapper;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import persistence.sql.ddl.constraint.ColumnH2Constraint;
+import persistence.sql.ddl.constraint.GeneratedValueH2Constraint;
+import persistence.sql.ddl.constraint.H2Constraint;
+import persistence.sql.ddl.constraint.IdH2Constraint;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
