+++ src/main/java/persistence/sql/ddl/mapper/H2TypeMapper.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.mapper;
 
 import jakarta.persistence.Column;
